requirements.txt
langchain>=0.1.0
openai>=1.0.0
boto3>=1.28.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
transformers>=4.30.0
torch>=2.0.0
tensorflow>=2.13.0
langgraph>=0.0.15
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0